수정, 디버그, size_t를 int 형식으로 형 변환

+++ main.cpp
@@ -126,7 +126,7 @@ bool BattlePhase(BattleManager& battleManager, MonsterPool& monsterPool)
             else
             {
                 std::vector<Monster*> monster = battleManager.GetLivingMonsters();
-                std::uniform_int_distribution<int> monsterDist(0, monster.size() - 1);
+                std::uniform_int_distribution<size_t> monsterDist(0, monster.size() - 1);
                 Monster* targetMonster = monster[monsterDist(gen)];
                 battleManager.PlayerHitMonster(targetMonster, turnPlayer->GetPower());
                 if (targetMonster->IsDead())
@@ -160,7 +160,7 @@ bool BattlePhase(BattleManager& battleManager, MonsterPool& monsterPool)
         {
             Monster* turnMonster = turnMonsters[i];
             std::vector<Player*> livingPlayers = battleManager.GetLivingPlayers();
-            std::uniform_int_distribution<int> monsterDist(0, livingPlayers.size() - 1);
+            std::uniform_int_distribution<size_t> monsterDist(0, livingPlayers.size() - 1);
             Player* targetPlayer = livingPlayers[monsterDist(gen)];
             battleManager.MonsterHitPlayer(targetPlayer, turnMonster->GetPower());
 
@@ -209,7 +209,7 @@ void NormalBattle()
         totalLv += players[i]->GetLevel();
     }
 
-    int avgLv = totalLv / players.size();
+    int avgLv = totalLv / (int)players.size();
 
     int monsterCount = std::max(1, avgLv / 2);
 
@@ -290,7 +290,7 @@ void NormalBattle()
     {
         totalLv += players[i]->GetLevel();
     }
-    avgLv = totalLv / players.size();
+    avgLv = totalLv / (int)players.size();
 
     while (true)
     {
@@ -332,7 +332,7 @@ void BossBattle()
         totalLv += players[i]->GetLevel();
     }
 
-    int avgLv = totalLv / players.size();
+    int avgLv = totalLv / (int)players.size();
 
     int monsterCount = 1;
 
@@ -413,7 +413,7 @@ void MainMenu()
     {
         totalLv += players[i]->GetLevel();
     }
-    int avgLv = totalLv / players.size();
+    int avgLv = totalLv / (int)players.size();
 
     // 상태 전이
     switch (option)
